Rename after typo on `sanitized`
