+++ backend/windmill-worker/src/lib.rs
@@ -41,7 +41,7 @@ mod rust_executor;
 mod worker;
 mod worker_flow;
 mod worker_lockfiles;
-mod santized_sql_params;
+mod sanitized_sql_params;
 mod schema;
 
 pub use worker::*;
